cluster: Split logic and improve the Astarte Resource build logic
Remove all hardwired logic from instance deploy, and split around utils/helpers all needed
hooks. This is a prerogative to implement all other commands which will act on profile reconciliation.
